package/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "apis"
authors = [
    {name = "Shreya Porwal",email = "comtact.porwalshreya@gmail.com"}
]

requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.132.0,<0.133.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "asyncmy (>=0.2.11,<0.3.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.poetry]
name = "server"
version = "0.1.0"
packages = [{ include = "app" }]

[tool.pytest.ini_options]
asyncio_mode = "auto"